requirements.txt
streamlit
langchain
langchain-community
langchain-groq
langgraph
duckduckgo-search
python-dotenv
ddgs